+++ dialects/sqlite-3-38/src/main/kotlin/app/cash/sqldelight/dialects/sqlite_3_38/SqliteTypeResolver.kt
@@ -3,14 +3,14 @@ package app.cash.sqldelight.dialects.sqlite_3_38
 import app.cash.sqldelight.dialect.api.IntermediateType
 import app.cash.sqldelight.dialect.api.PrimitiveType.TEXT
 import app.cash.sqldelight.dialect.api.TypeResolver
-import app.cash.sqldelight.dialects.sqlite_3_35.SqliteTypeResolver as Sqlite335TypeResolver
+import app.cash.sqldelight.dialects.sqlite_3_37.SqliteTypeResolver as Sqlite337TypeResolver
 import app.cash.sqldelight.dialects.sqlite_3_38.grammar.psi.SqliteExtensionExpr
 import com.alecstrong.sql.psi.core.psi.SqlExpr
 /*
-This class extends 3_35 SqliteTypeResolver as we need to call the inherited resolvers for previous dialects
+This class extends 3_37 SqliteTypeResolver as we need to call the inherited resolvers for previous dialects
 Supports type resolution for path operators (->, ->>) in result columns
  */
-open class SqliteTypeResolver(parentResolver: TypeResolver) : Sqlite335TypeResolver(parentResolver) {
+open class SqliteTypeResolver(parentResolver: TypeResolver) : Sqlite337TypeResolver(parentResolver) {
 
   override fun resolvedType(expr: SqlExpr): IntermediateType {
     return if (expr is SqliteExtensionExpr) {