+++ sqldelight-compiler/src/test/kotlin/app/cash/sqldelight/core/TestDialect.kt
@@ -12,6 +12,7 @@ enum class TestDialect(val dialect: SqlDelightDialect) {
   SQLITE_3_25(app.cash.sqldelight.dialects.sqlite_3_25.SqliteDialect()),
   SQLITE_3_30(app.cash.sqldelight.dialects.sqlite_3_30.SqliteDialect()),
   SQLITE_3_35(app.cash.sqldelight.dialects.sqlite_3_35.SqliteDialect()),
+  SQLITE_3_37(app.cash.sqldelight.dialects.sqlite_3_37.SqliteDialect()),
   SQLITE_3_38(app.cash.sqldelight.dialects.sqlite_3_38.SqliteDialect()),
   SQLITE_3_44(app.cash.sqldelight.dialects.sqlite_3_44.SqliteDialect()),
   POSTGRESQL(PostgreSqlDialect()),